requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
breadability==0.1.20
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
gensim==4.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.2.2
lxml_html_clean==0.1.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
priority==2.0.0
prompt_toolkit==3.0.47
pure_eval==0.2.3
pycountry==24.6.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
Quart==0.19.6
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rpds-py==0.19.1
scipy==1.13.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==7.0.4
soupsieve==2.5
stack-data==0.6.3
sumy==0.11.0
tinycss2==1.3.0
tinysegmenter==0.3
tldextract==5.1.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4